AI_Executive_Secretary/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.8.1
asyncpg==0.31.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distlib==0.3.9
distro==1.9.0
Django==5.1.6
djangorestframework==3.16.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.2
filelock==3.17.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
holidays==0.93
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
packaging==24.2
passlib==1.7.4
pipenv==2024.4.1
platformdirs==4.3.6
proto-plus==1.27.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
requests==2.33.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==20.29.3
virtualenvwrapper-win==1.2.7
